chore: hooks 분리 및 컴포넌트 이름 변경
